+++ bluesky/navdb/__init__.py
@@ -1,0 +1,1 @@
+from navdb import Navdatabase
