+++ bluesky/tools/loaddata.py
@@ -1,0 +1,49 @@
+import os
+try:
+    import cPickle as pickle
+except ImportError:
+    import pickle
+
+from load_visuals_txt import load_coastline_txt, load_aptsurface_txt
+from ..settings import data_path
+cachedir = data_path + '/cache'
+
+if not os.path.exists(cachedir):
+    os.makedirs(cachedir)
+
+
+def load_coastlines():
+    if not os.path.isfile(cachedir + '/coastlines.p'):
+        coastvertices, coastindices = load_coastline_txt()
+        with open(cachedir + '/coastlines.p', 'wb') as f:
+            pickle.dump(coastvertices, f, pickle.HIGHEST_PROTOCOL)
+            pickle.dump(coastindices, f, pickle.HIGHEST_PROTOCOL)
+    else:
+        with open(cachedir + '/coastlines.p', 'rb') as f:
+            coastvertices = pickle.load(f)
+            coastindices  = pickle.load(f)
+
+    return coastvertices, coastindices
+
+
+def load_aptsurface():
+    if not os.path.isfile(cachedir + '/aptsurface.p'):
+        vbuf_asphalt, vbuf_concrete, vbuf_runways, apt_ctr_lat, apt_ctr_lon, \
+            apt_indices = load_aptsurface_txt()
+        with open(cachedir + '/aptsurface.p', 'wb') as f:
+            pickle.dump(vbuf_asphalt, f, pickle.HIGHEST_PROTOCOL)
+            pickle.dump(vbuf_concrete, f, pickle.HIGHEST_PROTOCOL)
+            pickle.dump(vbuf_runways , f, pickle.HIGHEST_PROTOCOL)
+            pickle.dump(apt_ctr_lat  , f, pickle.HIGHEST_PROTOCOL)
+            pickle.dump(apt_ctr_lon  , f, pickle.HIGHEST_PROTOCOL)
+            pickle.dump(apt_indices  , f, pickle.HIGHEST_PROTOCOL)
+    else:
+        with open(cachedir + '/aptsurface.p', 'rb') as f:
+            vbuf_asphalt  = pickle.load(f)
+            vbuf_concrete = pickle.load(f)
+            vbuf_runways  = pickle.load(f)
+            apt_ctr_lat   = pickle.load(f)
+            apt_ctr_lon   = pickle.load(f)
+            apt_indices   = pickle.load(f)
+    return vbuf_asphalt, vbuf_concrete, vbuf_runways, apt_ctr_lat, \
+        apt_ctr_lon, apt_indices
