+++ bluesky/traf/__init__.py
@@ -1,2 +1,1 @@
 from traffic import Traffic
-from navdb import Navdatabase
